// frontend/StockManager/generated-sources/openapi/models/PortfolioCreateDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * StockManager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PortfolioCreateDto
 */
export interface PortfolioCreateDto {
    /**
     * 
     * @type {string}
     * @memberof PortfolioCreateDto
     */
    name?: string | null;
}

/**
 * Check if a given object implements the PortfolioCreateDto interface.
 */
export function instanceOfPortfolioCreateDto(value: object): value is PortfolioCreateDto {
    return true;
}

export function PortfolioCreateDtoFromJSON(json: any): PortfolioCreateDto {
    return PortfolioCreateDtoFromJSONTyped(json, false);
}

export function PortfolioCreateDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PortfolioCreateDto {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
    };
}

export function PortfolioCreateDtoToJSON(json: any): PortfolioCreateDto {
    return PortfolioCreateDtoToJSONTyped(json, false);
}

export function PortfolioCreateDtoToJSONTyped(value?: PortfolioCreateDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
    };
}