// frontend/StockManager/generated-sources/openapi/models/PortfolioUpdateDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * StockManager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PortfolioUpdateDto
 */
export interface PortfolioUpdateDto {
    /**
     * 
     * @type {string}
     * @memberof PortfolioUpdateDto
     */
    name?: string | null;
}

/**
 * Check if a given object implements the PortfolioUpdateDto interface.
 */
export function instanceOfPortfolioUpdateDto(value: object): value is PortfolioUpdateDto {
    return true;
}

export function PortfolioUpdateDtoFromJSON(json: any): PortfolioUpdateDto {
    return PortfolioUpdateDtoFromJSONTyped(json, false);
}

export function PortfolioUpdateDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PortfolioUpdateDto {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
    };
}

export function PortfolioUpdateDtoToJSON(json: any): PortfolioUpdateDto {
    return PortfolioUpdateDtoToJSONTyped(json, false);
}

export function PortfolioUpdateDtoToJSONTyped(value?: PortfolioUpdateDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
    };
}